reimplemented heartbeats
We discovered that RabbitMQ would hang up on consumers a few times
a day because of heartbeat timeouts.  Time/clocks/et al are known
to be flaky in containers and we observed some time.sleep() calls
exceeding their target by x3.

Instead of using a child thread to send heartbeats, we use the
child thread to process messages.  The main thread now sends
heartbeats.  Furthermore, the heartbeater class has been retired
and the heartbeat is implemented in the idle loop (using a socket
timeout) and in the busy loop (timeout waiting for the thread
to finish processing the message).  As a final defense, we
send heartbeats at a more frequent interval than necessary.

Testing has shown that this is much more robust than before.

refs thugs-1836