Handle TERM signal and set a flag to terminate
As we run this in Kubernetes we do not want to terminate immediately, nor do we want to ignore the TERM and continue on. We want to set a flag for termination and make sure any in-flight message may continue processing until completion. Before a new message starts processing, we will raise SystemExit.